+++ packages/uploader/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.97 (June 24, 2025)
+
+- fix: resolve the bug of upload video error
+
+## 0.1.96 (June 24, 2025)
+
+fix: fix image paste upload issue in Safari
+
 ## 0.1.95 (May 31, 2025)
 
 - feat: improve virtual-plugin in uploader